Finalizei o método de deletar Tarefas e fiz algumas correções porque o código não estava rodando

+++ task-maneger-console-springBoot/src/main/java/com/br/task_maneger_console_springBoot/service/TarefaService.java
@@ -56,9 +56,20 @@ public class TarefaService {
     }
 
 
-    public void deletarTarefa(Long id){
+    public boolean deletarTarefa(Long id){
 
-        tarefaRepository.deleteById(id);
+        if (tarefaRepository.existsById(id)){
+
+            tarefaRepository.deleteById(id);
+
+            return true;
+
+        }else{
+
+            return false;
+
+        }
+
 
     }
